refactor: Put into cart directory if it is related to cart components
