+++ frontend/src/components/molecules/cart/cartColumn/CartColumn.scss
