+++ frontend/src/components/molecules/cart/cartSummary/CartSummary.scss
