+++ frontend/src/components/molecules/index.ts
@@ -5,11 +5,9 @@ export * from './itemColumn';
 export * from './button';
 export * from './object';
 export * from './form';
-export * from './cartColumn';
-export * from './cartSummary';
-export * from './cartAddress';
 export * from './orderColumn';
 export * from './productImage';
+export * from './cart';
 
 // export * from './button';
 // export * from './intro';
